+++ app/controllers/computations_controller.rb
@@ -9,24 +9,21 @@ class ComputationsController < ApplicationController
   end
 
   def create
-    @computation = Computation.create(
-      create_params.merge(
-        user: current_user,
-        computation_type: Computation.type_for_patient_status(patient.procedure_status),
-        script: ComputationScriptGenerator.new(patient, current_user).script
-      )
-    )
-    Rimrock::StartJob.perform_later @computation
+    @computation = computation_builder_for(patient.procedure_status).run(patient, current_user)
     redirect_to @computation.patient, notice: 'Computation submitted'
   end
 
   private
+
+  def computation_builder_for(procedure_status)
+    Patient::PIPELINE[procedure_status.to_sym]
+  end
 
   def create_params
     params.require(:computation).permit(:patient_id)
   end
 
   def patient
-    @patient ||= Patient.find(params[:computation][:patient_id])
+    @patient ||= Patient.find(create_params[:patient_id])
   end
 end