+++ app/models/computation.rb
@@ -1,19 +1,14 @@
 # frozen_string_literal: true
 class Computation < ApplicationRecord
-  enum computation_type: [
-    :blood_flow_simulation,
-    :heart_model_computation
-  ]
-
   belongs_to :user
   belongs_to :patient
 
-  validates :script, presence: true
   validates :user, presence: true
   validates :status, inclusion: { in: %w(new queued running error finished aborted) }
+  validates :pipeline_step, inclusion: { in: Patient::PIPELINE.keys.map(&:to_s) }
 
   scope :active, -> { where(status: %w(new queued running)) }
-  scope :for_patient_status, ->(status) { where(computation_type: type_for_patient_status(status)) }
+  scope :for_patient_status, ->(status) { where(pipeline_step: status) }
 
   def active?
     %w(new queued running).include? status
@@ -21,6 +16,7 @@ class Computation < ApplicationRecord
 
   def self.type_for_patient_status(status)
     case status
+    when 'imaging_uploaded' then 'segmentation'
     when 'virtual_model_ready' then 'blood_flow_simulation'
     when 'after_parameter_estimation' then 'heart_model_computation'
     end