+++ app/models/pipeline/segmentation.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+module Pipeline
+  class Segmentation
+  end
+end