+++ db/migrate/20170308121133_remove_computation_type_from_computations.rb
@@ -1,0 +1,5 @@
+class RemoveComputationTypeFromComputations < ActiveRecord::Migration[5.0]
+  def change
+    remove_column :computations, :computation_type, :string
+  end
+end