+++ db/migrate/20170308121238_add_type_to_computations_for_sti.rb
@@ -1,0 +1,5 @@
+class AddTypeToComputationsForSti < ActiveRecord::Migration[5.0]
+  def change
+    add_column :computations, :type, :string
+  end
+end