+++ db/migrate/20170308145040_add_output_path_to_computations.rb
@@ -1,0 +1,5 @@
+class AddOutputPathToComputations < ActiveRecord::Migration[5.0]
+  def change
+    add_column :computations, :output_path, :string
+  end
+end