+++ db/migrate/20170313131252_add_pipeline_step_to_computations.rb
@@ -1,0 +1,5 @@
+class AddPipelineStepToComputations < ActiveRecord::Migration[5.0]
+  def change
+    add_column :computations, :pipeline_step, :string
+  end
+end