+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20170227093146) do
+ActiveRecord::Schema.define(version: 20170313131252) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -52,7 +52,10 @@ ActiveRecord::Schema.define(version: 20170227093146) do
     t.datetime "created_at",                        null: false
     t.datetime "updated_at",                        null: false
     t.integer  "patient_id"
-    t.integer  "computation_type",  default: 0,     null: false
+    t.string   "type"
+    t.string   "input_path"
+    t.string   "output_path"
+    t.string   "pipeline_step"
     t.index ["patient_id"], name: "index_computations_on_patient_id", using: :btree
   end