+++ spec/models/computation_spec.rb
@@ -4,11 +4,15 @@ require 'rails_helper'
 RSpec.describe Computation, type: :model do
   subject { create(:computation) }
 
-  it { should validate_presence_of(:script) }
   it { should validate_presence_of(:user) }
   it do
     should validate_inclusion_of(:status).
       in_array(%w(new queued running error finished aborted))
+  end
+
+  it do
+    should validate_inclusion_of(:pipeline_step).
+      in_array(Patient::PIPELINE.keys.map(&:to_s))
   end
 
   it { should belong_to(:user) }
@@ -22,15 +26,6 @@ RSpec.describe Computation, type: :model do
       expect(Computation.active).to eq [subject]
       subject.update(status: 'finished')
       expect(Computation.active).to be_empty
-    end
-  end
-
-  describe '.type_for_patient_status' do
-    it 'properly maps patient procedure status to required computation' do
-      expect(Computation.type_for_patient_status('virtual_model_ready')).
-        to eq 'blood_flow_simulation'
-      expect(Computation.type_for_patient_status('after_parameter_estimation')).
-        to eq 'heart_model_computation'
     end
   end
 end