+++ spec/models/pipeline/rimrock_based_step_shared_examples.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: true
+require 'rails_helper'
+
+shared_examples 'a Rimrock-based step' do
+  before do
+    allow(Rimrock::StartJob).to receive(:perform_later)
+  end
+
+  it 'creates RimrockComputation' do
+    expect { described_class.run(patient, user) }.
+      to change { RimrockComputation.count }.by(1)
+  end
+
+  it 'returns a RimrockComputation object' do
+    computation = described_class.run(patient, user)
+    expect(computation.class).to eq RimrockComputation
+  end
+
+  it 'starts a Rimrock job' do
+    expect(Rimrock::StartJob).to receive(:perform_later)
+    described_class.run(patient, user)
+  end
+end